had to switcheroo rail_fence_cipher.c and example.c
